+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `list-context.image-list` has been refactored to use the `image-list` with a `list-context`
 
 ### Added
-- `image-list` is now a simple list of images that can be editable from the admin panel 
+- `image-list` a new block which renders a simple list of images that can be editable via the Site Editor 
 
 ## [0.5.3] - 2020-10-07
 ### Removed